Create README.md
I am a junior consultant to an ISP and will like to develop myself very rapidly on coding in php and python. I have a fair working knowledge of Java and C but will like to sharpen my skill more in Java to be able to handle cryptography applications, since that's my area of interest and specialisation.